+++ lib/storage/storage.go
@@ -276,7 +276,7 @@ func OpenStorage(path string, retentionMsecs int64, maxHourlySeries, maxDailySer
 
 var maxTSIDCacheSize int
 
-// SetTSIDCacheSize overrides the default size of storage/tsid cahce
+// SetTSIDCacheSize overrides the default size of storage/tsid cache
 func SetTSIDCacheSize(size int) {
 	maxTSIDCacheSize = size
 }